+++ tests/plugins/environments/test_lightning_environment.py
@@ -14,6 +14,8 @@
 import os
 from unittest import mock
 
+import pytest
+
 from pytorch_lightning.plugins.environments import LightningEnvironment
 
 
@@ -48,6 +50,20 @@ def test_attributes_from_environment_variables():
     assert env.global_rank() == 100
     env.set_world_size(100)
     assert env.world_size() == 100
+
+
+@pytest.mark.parametrize(
+    "environ, creates_children", [
+        ({}, False),
+        (dict(LOCAL_RANK="2"), True),
+        (dict(NODE_RANK="1"), False),
+    ]
+)
+def test_manual_user_launch(environ, creates_children):
+    """ Test that the environment switches to manual user mode when LOCAL_RANK env variable detected. """
+    with mock.patch.dict(os.environ, environ):
+        env = LightningEnvironment()
+        assert env.creates_children() == creates_children
 
 
 @mock.patch.dict(os.environ, {